refactor(#112): split _normalize.py into cohesive subsystems
Extract _normalize.py (3179 lines) into four focused modules:

- _normalize_ir.py (817 lines): StateTemplate dataclass, shaped-param
  scanning, time-varying stripping, alias/equation IR builders, and
  string derivation helpers (_derive_equation_strings, _derive_alias_strings).

- _normalize_kernels.py (777 lines): kernel, operator, state-axes, and
  apply_along normalization (_ALLOWED_KERNEL_FORMS, _normalize_kernels,
  _normalize_operators, _select_apply_along_kernel, etc.).

- _normalize_initial_state.py (218 lines): initial_state block expansion
  (scalar and shaped forms).

- _normalize_chains.py (487 lines): chain helper expansion and coord-shift
  expansion.

_normalize.py is trimmed to 848 lines containing only the public entry
points (normalize_rhs, normalize_expr_rhs, normalize_transitions_rhs),
the two RHS dataclasses (ExprRhs, TransitionsRhs), and the thin
_normalize_common_meta / _build_transition_equations_ir helpers.

Also:
- Move _get_required_str into _helpers.py (shared by kernels + chains).
- Update _ir_expand.py late import to target _normalize_kernels instead
  of _normalize (the circular-import reason for the late import still
  applies, but the cycle is now _normalize_ir → _ir_expand → _normalize_kernels).
- Update test_op_system_specs.py to import _derive_* from _normalize_ir.

378 tests pass.
